refactor(disc): one canonical discipline source across both engines; v0.3.335
DISC-3b. New aec_data/disciplines.py holds the shared low-level discipline data —
MF_DIVISIONS (CSI MasterFormat division master) + the discipline colour palette
(DISCIPLINE_COLORS/SERIES_COLORS/discipline_color/division_of). It lives in aec_data
because aec_api can import aec_data but not vice-versa.

- aec_api.classification imports these instead of defining its own copies.
- aec_data.specmanual drops its duplicate _DIVISIONS map (was near-identical to
  MF_DIVISIONS, only "&"/"and" title differences) -> uses the canonical table.

So the project manual's CSI divisions and the estimate/sheet/viewer vocabularies are
literally the same objects. test_classification points colour asserts at the canonical
source. Verified: MF_DIVISIONS is now shared (identity); app imports clean; ruff clean;
9 classification-consumer test suites pass (discipline/naming/drawing/sheetgen/reports/
lod/mep/disciplines/code_analysis).
