+++ apps/web/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@aec/web",
-  "version": "0.3.334",
+  "version": "0.3.335",
   "private": true,
   "type": "module",
   "engines": {
